+++ playgrounds/gregor/src/main/java/playground/gregor/sim2d_v4/simulation/ORCAAgentFactory.java
@@ -30,7 +30,7 @@ import playground.gregor.sim2d_v4.simulation.physics.Sim2DAgent;
 public class ORCAAgentFactory implements Sim2DAgentFactory {
 
 	@Override
-	public Sim2DAgent buildAgent(QVehicle veh, float spawnX, float spawnY) {
+	public Sim2DAgent buildAgent(QVehicle veh, double spawnX, double spawnY) {
 		DelegableSim2DAgent delegate = new ORCAAgent(veh, spawnX, spawnY);
 		Sim2DAgent agent = new FailsafeAgentImpl(delegate);
 		return delegate;
